# docs/api/turn-2-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/businesses": {
            "get": {
                "tags": [
                    "Businesses"
                ],
                "summary": "List the businesses the caller belongs to",
                "operationId": "listMine",
                "responses": {
                    "200": {
                        "description": "The caller's businesses, possibly empty",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BusinessResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Create a business; the caller becomes its owner",
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateBusinessRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation failed or unknown timezone",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/businesses/{businessId}/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "List this business's services",
                "operationId": "list",
                "parameters": [
                    {
                        "name": "businessId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ServiceResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Add a service this business offers",
                "operationId": "create_1",
                "parameters": [
                    {
                        "name": "businessId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateService"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation failed",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not a member of this business",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "A service of that name exists",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/businesses/{businessId}": {
            "get": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Read one business the caller belongs to",
                "operationId": "get",
                "parameters": [
                    {
                        "name": "businessId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The business",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not a member \u2014 returned identically when the business does not exist",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/businesses/{businessId}/services/{serviceId}": {
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Remove a service and its employee links",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "businessId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "serviceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removed"
                    },
                    "403": {
                        "description": "Not a member of this business"
                    },
                    "404": {
                        "description": "No such service here"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateBusinessRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 120,
                        "minLength": 0
                    },
                    "timezone": {
                        "type": "string",
                        "maxLength": 64,
                        "minLength": 0
                    }
                }
            },
            "BusinessResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "timezone": {
                        "type": "string"
                    }
                }
            },
            "CreateService": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 120,
                        "minLength": 0
                    },
                    "durationMinutes": {
                        "type": "integer",
                        "format": "int32",
                        "maximum": 1440,
                        "minimum": 1
                    },
                    "priceMinor": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "ServiceResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "durationMinutes": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "priceMinor": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}